Structurize project according to the style guide
Since structuring of a project makes it a lot easier to navigate through project files, current project files were re-arranged to keep them in line with the following style guide - http://docs.python-guide.org/en/latest/writing/structure/ .
