+++ python/README.txt
@@ -1,0 +1,3 @@
+This is a template project containing training materials for Python automation.
+
+For more info please check ./docs directory.
