+++ python/docs/git.txt
@@ -1,0 +1,11 @@
+GIT
+
+mkdir /root/python/practice
+cd /root/python/practice
+git clone https://github.com/belgeu/misc.git
+cd misc
+## edit .git/config to ask pass (add belgeu@ to url)
+## url = https://belgeu@github.com/belgeu/misc.git
+git add webdriver_test.py
+git commit -m "MESSAGE"
+git push origin master
