+++ python/pip_packages.txt
@@ -1,0 +1,2 @@
+###### PIP packages ######
+selenium
